refactor(step2): 抽象 DocumentSymbol 构建工具，统一两个 provider 的 Range 计算
新增 src/common/symbols.ts：
- SymbolEntry 接口 — 通用的符号条目类型
- buildDocumentSymbols() — 从条目列表构建 vscode.DocumentSymbol[]

CSMLogDocumentSymbolProvider 和 LvcsmDocumentSymbolProvider 均改为使用
buildDocumentSymbols()，消除了 ~30 行重复的 Range 计算逻辑。

所有 79 个单元测试通过。
